+++ DAL.MySql.Schema/Tables/usersdata.sql
@@ -4,6 +4,6 @@ CREATE TABLE `usersdata` (
   `UserId` bigint unsigned NOT NULL,
   `GuildId` bigint unsigned DEFAULT NULL,
   `MeasuredOn` datetime(6) NOT NULL,
-  `Type` int NOT NULL,
+  `Type` tinyint unsigned NOT NULL,
   PRIMARY KEY (`Id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci
